object_detection/cocoEval_requirements.txt
matplotlib==3.2.2
numpy==1.19.5
skimage==0.16.2
json==2.0.9
pycocotools
pylab-sdk